+++ frontend/src/wailsjs/wailsjs/go/models.ts
@@ -139,6 +139,60 @@ export namespace store {
 	        this.scrollSpeedDown = source["scrollSpeedDown"];
 	    }
 	}
+	export class MidiMapping {
+	    type: string;
+	    number: number;
+	    channel: number;
+	
+	    static createFrom(source: any = {}) {
+	        return new MidiMapping(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.type = source["type"];
+	        this.number = source["number"];
+	        this.channel = source["channel"];
+	    }
+	}
+	export class MidiSettings {
+	    enabled: boolean;
+	    scrollDown?: MidiMapping;
+	    scrollUp?: MidiMapping;
+	    playPause?: MidiMapping;
+	    expressionScroll?: MidiMapping;
+	
+	    static createFrom(source: any = {}) {
+	        return new MidiSettings(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.enabled = source["enabled"];
+	        this.scrollDown = this.convertValues(source["scrollDown"], MidiMapping);
+	        this.scrollUp = this.convertValues(source["scrollUp"], MidiMapping);
+	        this.playPause = this.convertValues(source["playPause"], MidiMapping);
+	        this.expressionScroll = this.convertValues(source["expressionScroll"], MidiMapping);
+	    }
+	
+		convertValues(a: any, classs: any, asMap: boolean = false): any {
+		    if (!a) {
+		        return a;
+		    }
+		    if (a.slice && a.map) {
+		        return (a as any[]).map(elem => this.convertValues(elem, classs));
+		    } else if ("object" === typeof a) {
+		        if (asMap) {
+		            for (const key of Object.keys(a)) {
+		                a[key] = new classs(a[key]);
+		            }
+		            return a;
+		        }
+		        return new classs(a);
+		    }
+		    return a;
+		}
+	}
 	export class RemoteFile {
 	    name: string;
 	    path: string;
@@ -171,6 +225,7 @@ export namespace store {
 	    autoSyncFrequency: string;
 	    lastSyncTime: number;
 	    keyBindings: KeyBindings;
+	    midiSettings: MidiSettings;
 	    storagePath: string;
 	    coversPath: string;
 	    webdavEnabled: boolean;
@@ -197,6 +252,7 @@ export namespace store {
 	        this.autoSyncFrequency = source["autoSyncFrequency"];
 	        this.lastSyncTime = source["lastSyncTime"];
 	        this.keyBindings = this.convertValues(source["keyBindings"], KeyBindings);
+	        this.midiSettings = this.convertValues(source["midiSettings"], MidiSettings);
 	        this.storagePath = source["storagePath"];
 	        this.coversPath = source["coversPath"];
 	        this.webdavEnabled = source["webdavEnabled"];